Updates the flow to create only the services before the deployment.
It also separates the push related code for link and service creation.
